Extract Azure related function as a separated deployment module
refs oursky/likecoin-chain#24
