+++ deploy/cmd/deploy/assets/liked.service.template
