+++ deploy/cmd/deploy/assets/node-setup.sh
